Feature/milestone 03

+++ src/main/java/ch/bzz/autohaus/service/AutoService.java
@@ -1,10 +1,13 @@
 package ch.bzz.autohaus.service;
 
+import ch.bzz.autohaus.assets.Helper;
 import ch.bzz.autohaus.data.DataHandler;
 import ch.bzz.autohaus.model.Auto;
+import ch.bzz.autohaus.model.User;
 
 import javax.validation.Valid;
 import javax.ws.rs.BeanParam;
+import javax.ws.rs.CookieParam;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
@@ -14,6 +17,7 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
 
@@ -28,16 +32,35 @@ public class AutoService {
     /**
      * Delivers autoList as a JsonArray
      *
-     * @return Response with Status OK and the autoList
+     * @param encryptedUsername encrypted username from cookie
+     * @param encryptedPassword encrypted password from cookie
+     * @return Response with Status 200, 401 or 403 and the autoList
      */
     @GET
     @Path("list")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response listAutos() {
-        List<Auto> autoList = DataHandler.getInstance().readAllAutos();
+    public Response listAutos(
+            @CookieParam("username") String encryptedUsername,
+            @CookieParam("password") String encryptedPassword
+    ) {
+        List<Auto> autoList = Collections.emptyList();
+        User loggedInUser = Helper.getInstance().getUserByEncryptedLogin(encryptedUsername, encryptedPassword);
+
+        int httpStatus = 200;
+
+        if (loggedInUser != null) {
+            if (Helper.getInstance().isUserValidForRead(loggedInUser)) {
+                autoList = DataHandler.getInstance().readAllAutos();
+            } else {
+                httpStatus = 403;
+            }
+        } else {
+            httpStatus = 401;
+        }
+
 
         return Response
-                .status(200)
+                .status(httpStatus)
                 .entity(autoList)
                 .build();
     }
@@ -45,25 +68,39 @@ public class AutoService {
     /**
      * Delivers an auto with a specific uuid
      *
-     * @param id uuid of the auto
-     * @return Response with Status 200, 400 or 404 (depends on if an entity could be found) and the auto
+     * @param id                uuid of the auto
+     * @param encryptedUsername encrypted username from cookie
+     * @param encryptedPassword encrypted password from cookie
+     * @return Response with Status 200, 400, 401, 403 or 404 (depends on if an entity could be found) and the auto
      */
     @GET
     @Path("auto")
     @Produces(MediaType.APPLICATION_JSON)
     public Response idAuto(
-            @QueryParam("id") String id
+            @QueryParam("id") String id,
+            @CookieParam("username") String encryptedUsername,
+            @CookieParam("password") String encryptedPassword
     ) {
+        User loggedInUser = Helper.getInstance().getUserByEncryptedLogin(encryptedUsername, encryptedPassword);
         Auto auto = null;
+
         int httpStatus = 200;
 
-        try {
-            auto = DataHandler.getInstance().readAutoByUUID(id);
-            if (auto == null) {
-                httpStatus = 404;
+        if (loggedInUser != null) {
+            if (Helper.getInstance().isUserValidForRead(loggedInUser)) {
+                try {
+                    auto = DataHandler.getInstance().readAutoByUUID(id);
+                    if (auto == null) {
+                        httpStatus = 404;
+                    }
+                } catch (Exception exception) {
+                    httpStatus = 400;
+                }
+            } else {
+                httpStatus = 403;
             }
-        } catch (Exception exception) {
-            httpStatus = 400;
+        } else {
+            httpStatus = 401;
         }
         return Response
                 .status(httpStatus)
@@ -74,23 +111,37 @@ public class AutoService {
     /**
      * Creates an auto
      *
-     * @param auto auto BeanParam
-     * @return Response with Status 200 or 400
+     * @param auto              auto BeanParam
+     * @param encryptedUsername encrypted username from cookie
+     * @param encryptedPassword encrypted password from cookie
+     * @return Response with Status 200, 400, 401 or 403
      */
     @POST
     @Path("create")
     @Produces(MediaType.TEXT_PLAIN)
     public Response createAuto(
-            @Valid @BeanParam Auto auto
+            @Valid @BeanParam Auto auto,
+            @CookieParam("username") String encryptedUsername,
+            @CookieParam("password") String encryptedPassword
     ) {
+        User loggedInUser = Helper.getInstance().getUserByEncryptedLogin(encryptedUsername, encryptedPassword);
+
         int httpStatus = 200;
 
-        auto.setAutoUUID(UUID.randomUUID().toString());
+        if (loggedInUser != null) {
+            if (Helper.getInstance().isUserValidForCreate(loggedInUser)) {
+                auto.setAutoUUID(UUID.randomUUID().toString());
 
-        try {
-            DataHandler.getInstance().insertAuto(auto);
-        } catch (Exception exception) {
-            httpStatus = 400;
+                try {
+                    DataHandler.getInstance().insertAuto(auto);
+                } catch (Exception exception) {
+                    httpStatus = 400;
+                }
+            } else {
+                httpStatus = 403;
+            }
+        } else {
+            httpStatus = 401;
         }
         return Response
                 .status(httpStatus)
@@ -101,28 +152,42 @@ public class AutoService {
     /**
      * updates an auto
      *
-     * @param auto auto BeanParam
-     * @return Response with Status 200, 400 or 410
+     * @param auto              auto BeanParam
+     * @param encryptedUsername encrypted username from cookie
+     * @param encryptedPassword encrypted password from cookie
+     * @return Response with Status 200, 400, 401, 403 or 410
      */
     @PUT
     @Path("update")
     @Produces(MediaType.TEXT_PLAIN)
     public Response updateAuto(
-            @Valid @BeanParam Auto auto
+            @Valid @BeanParam Auto auto,
+            @CookieParam("username") String encryptedUsername,
+            @CookieParam("password") String encryptedPassword
     ) {
+        User loggedInUser = Helper.getInstance().getUserByEncryptedLogin(encryptedUsername, encryptedPassword);
+
         int httpStatus = 200;
 
-        try {
-            Auto autoToBeUpdated = DataHandler.getInstance().readAutoByUUID(auto.getAutoUUID());
-            if (autoToBeUpdated != null) {
-                setAttributes(autoToBeUpdated, auto);
+        if (loggedInUser != null) {
+            if (Helper.getInstance().isUserValidForUpdate(loggedInUser)) {
+                try {
+                    Auto autoToBeUpdated = DataHandler.getInstance().readAutoByUUID(auto.getAutoUUID());
+                    if (autoToBeUpdated != null) {
+                        setAttributes(autoToBeUpdated, auto);
 
-                DataHandler.getInstance().updateAuto();
+                        DataHandler.getInstance().updateAuto();
+                    } else {
+                        httpStatus = 410;
+                    }
+                } catch (Exception exception) {
+                    httpStatus = 400;
+                }
             } else {
-                httpStatus = 410;
+                httpStatus = 403;
             }
-        } catch (Exception exception) {
-            httpStatus = 400;
+        } else {
+            httpStatus = 401;
         }
         return Response
                 .status(httpStatus)
@@ -133,23 +198,37 @@ public class AutoService {
     /**
      * Deletes an auto identified by its uuid
      *
-     * @param id uuid of the auto
+     * @param id                uuid of the auto
+     * @param encryptedUsername encrypted username from cookie
+     * @param encryptedPassword encrypted password from cookie
      * @return Response
      */
     @DELETE
     @Path("delete")
     @Produces(MediaType.TEXT_PLAIN)
     public Response deleteAuto(
-            @QueryParam("id") String id
+            @QueryParam("id") String id,
+            @CookieParam("username") String encryptedUsername,
+            @CookieParam("password") String encryptedPassword
     ) {
+        User loggedInUser = Helper.getInstance().getUserByEncryptedLogin(encryptedUsername, encryptedPassword);
+
         int httpStatus = 200;
 
-        try {
-            if (!DataHandler.getInstance().deleteAuto(id)) {
-                httpStatus = 410;
+        if (loggedInUser != null) {
+            if (Helper.getInstance().isUserValidForDelete(loggedInUser)) {
+                try {
+                    if (!DataHandler.getInstance().deleteAuto(id)) {
+                        httpStatus = 410;
+                    }
+                } catch (Exception exception) {
+                    httpStatus = 400;
+                }
+            } else {
+                httpStatus = 403;
             }
-        } catch (Exception exception) {
-            httpStatus = 400;
+        } else {
+            httpStatus = 401;
         }
         return Response
                 .status(httpStatus)
